Rename rest package to v1alpha1.
