+++ pkg/visibility/api/v1alpha1/cluster_queue.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rest
+package v1alpha1
 
 import (
 	"k8s.io/apimachinery/pkg/runtime"
