+++ pkg/visibility/api/v1alpha1/pending_workloads_cq_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rest
+package v1alpha1
 
 import (
 	"context"
