+++ pkg/visibility/api/v1alpha1/test_utils.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rest
+package v1alpha1
 
 import (
 	visibility "sigs.k8s.io/kueue/apis/visibility/v1alpha1"
